requirements.txt
Django==2.1b1
ipython==6.4.0
remote-pdb==1.2.0
requests==2.19.1
powerline-status==2.6
celery==4.2.0
django-celery-results==1.0.1
flower==0.9.2
psycopg2-binary==2.7.5
python-dateutil==2.7.3
numpy==1.15.0rc1
pandas==0.23.1
lxml==4.2.3
gevent==1.3.5
grequests==0.3.0
python-telegram-bot==10.1.0
pysocks==1.6.8

pytest==3.6.3
pytest-runner==4.2
parse==1.8.4
six==1.11.0
coverage==4.5.1
watchdog==0.8.3
tox==3.1.1
flake8==3.5.0